Feature/adjust security to application

+++ src/main/java/pl/orange/NextDoorBook/book/BookController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 import pl.orange.NextDoorBook.book.dto.BookAddDTO;
 import pl.orange.NextDoorBook.book.dto.BookDTO;
 import pl.orange.NextDoorBook.book.dto.BookInfoDTO;
+import pl.orange.NextDoorBook.security.payloads.MessageResponse;
 import pl.orange.NextDoorBook.user.User;
 
 import java.util.List;
@@ -35,8 +36,13 @@ public class BookController {
 
     @DeleteMapping(path = "/{id}")
     @PreAuthorize("hasRole('USER')")
-    public ResponseEntity deleteBook(@PathVariable Long id, Authentication authentication) {
-        Object principal = authentication.getPrincipal();
+    public ResponseEntity<?> deleteBook(@PathVariable Long id, UsernamePasswordAuthenticationToken user) {
+        User userFromObjectMapper = objectMapper.convertValue(user.getPrincipal(), User.class);
+
+        if (!bookService.isBookOwnedByUser(id,userFromObjectMapper)){
+            return ResponseEntity.badRequest().body(new MessageResponse("Error: Book has another owner"));
+        }
+
         bookService.deleteBook(id);
         return ResponseEntity
                 .status(200)
@@ -117,7 +123,13 @@ public class BookController {
 
     @PatchMapping(path = "")
     @PreAuthorize("hasRole('USER')")
-    public ResponseEntity<BookDTO> updateBook(@RequestBody BookDTO requestBook) {
+    public ResponseEntity<?> updateBook(@RequestBody BookDTO requestBook
+            , UsernamePasswordAuthenticationToken user) {
+
+        User userFromObjectMapper = objectMapper.convertValue(user.getPrincipal(), User.class);
+        if (!bookService.isBookOwnedByUser(requestBook.id(), userFromObjectMapper)){
+            return ResponseEntity.badRequest().body(new MessageResponse("Error: Book has another owner"));
+        }
         return ResponseEntity
                 .status(200)
                 .body(bookService.updateBook(requestBook));
@@ -126,6 +138,7 @@ public class BookController {
     @GetMapping("info/{bookId}")
     @PreAuthorize("hasRole('USER')")
     public ResponseEntity<BookInfoDTO> getBookInfo(@PathVariable Long bookId) {
+
         return ResponseEntity
                 .status(200)
                 .body(bookService.getBookInfo(bookId));
